Added contribution guidelines for simulation method developers!
Added MyNewBackend to all files where necessary so that it will be easier to find for developers.

Changed the way to install submodules so that they are now in the requirements.txt as 'editable' packages. This way we don't have to re-install. Also added these folders to .vscode/settings.json to make the intellisense (PyLance) work with these now-local packages.

Added debug_celery flag for easier switching between celery and debugging.

Added ids to the example-settings/examples.json file to have more correct examples for parameters.

Finally, I updated the .gitmodules file to have http rather than ssh access to the submodules. Will have to figure out the best way to do this.